@@ -10,17 +10,19 @@
   },
   "dependencies": {
     "@types/react": "18.0.37",
+    "autoprefixer": "^10.4.14",
     "axios": "^1.3.6",
     "bootstrap": "^5.2.3",
     "cookies-next": "^2.1.1",
     "eslint": "8.38.0",
     "eslint-config-next": "13.3.0",
     "moment": "^2.29.4",
-    "next": "13.3.0",
+    "next": "^13.4.3",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
     "react-dom": "18.2.0",
     "react-icons": "^4.8.0",
+    "tailwindcss": "^3.3.2",
     "typescript": "5.0.4"
   }
 }
